update go install script
* allow providing VERSION via ENV
* always delete pkg
* check if version is already extracted
* do not treat 1.5x special

+++ src/articles/setup-go/install.sh
@@ -1,19 +1,26 @@
 #!/bin/sh
 set -xe
 
-VERSION=1.4.2
+VERSION=${VERSION-:1.5.1}
 ARCH=$(uname | awk '{print tolower($0)}')-amd64
 DST=/usr/local/go${VERSION}
 
-if ! which curl; then
-  apt-get install -y curl
+echo "using VERSION=${VERSION}"
+
+if [[ ! -e $DST/bin/go ]]; then
+	if ! which curl; then
+		apt-get install -y curl
+	fi
+
+	TMP=$(mktemp -d /tmp/goXXXXXX)
+	cd ${TMP} && curl -SfL https://storage.googleapis.com/golang/go${VERSION}.${ARCH}.tar.gz | tar xfz - --strip=1
+
+	sudo mv ${TMP} ${DST}
+	sudo chmod 0755 ${DST}
 fi
 
-TMP=$(mktemp -d /tmp/goXXXXXX)
-cd ${TMP} && curl -SfL https://storage.googleapis.com/golang/go${VERSION}.${ARCH}.tar.gz | tar xfz - --strip=1
+rm -Rf $HOME/pkg
 
-sudo mv ${TMP} ${DST}
-sudo chmod 0755 ${DST}
 sudo ln -nfs ${DST} /usr/local/go
 
 # it is also helpful to delete $GOPATH/pkg (at least when upgrading go < 1.5)
